Move test folder inside src
