+++ packages/azure-store/tsconfig.json
@@ -2,8 +2,9 @@
   "$schema": "https://json.schemastore.org/tsconfig.json",
   "extends": "../../tsconfig.base.json",
   "references": [{ "path": "../utils/tsconfig.json" }],
-  "include": ["src", "test"],
+  "include": ["src"],
   "compilerOptions": {
+    "rootDir": "src",
     "outDir": "dist",
   }
 }
