+++ packages/server/src/test/PostHandler.test.ts
@@ -8,7 +8,7 @@ import httpMocks from 'node-mocks-http'
 import sinon from 'sinon'
 
 import {EVENTS, Upload, DataStore, type CancellationContext} from '@tus/utils'
-import {PostHandler} from '../src/handlers/PostHandler.js'
+import {PostHandler} from '../handlers/PostHandler.js'
 import {addPipableStreamBody} from './utils.js'
 import {MemoryLocker} from '@tus/server'
