+++ packages/server/tsconfig.json
@@ -1,9 +1,13 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig.json",
   "extends": "../../tsconfig.base.json",
-  "references": [{ "path": "../utils/tsconfig.json" }],
-  "include": ["src", "test"],
+  "references": [
+    { "path": "../utils/tsconfig.json" },
+    { "path": "../file-store/tsconfig.json" }
+  ],
+  "include": ["src"],
   "compilerOptions": {
+    "rootDir": "src",
     "outDir": "dist",
   }
 }
