+++ packages/utils/package.json
@@ -3,7 +3,8 @@
   "name": "@tus/utils",
   "version": "0.5.1",
   "description": "Internal utils for tus Node.js server and stores",
-  "main": "dist/index.js",
+  "main": "./dist/index.js",
+  "exports": "./dist/index.js",
   "type": "module",
   "homepage": "https://github.com/tus/tus-node-server#readme",
   "bugs": "https://github.com/tus/tus-node-server/issues",
