+++ packages/utils/src/test/Upload.test.ts
@@ -1,8 +1,8 @@
 import 'should'
 import {strict as assert} from 'node:assert'
 
-import {Upload} from '../src/models/Upload.js'
-import {Uid} from '../src/models/Uid.js'
+import {Upload} from '../models/Upload.js'
+import {Uid} from '../models/Uid.js'
 
 describe('Upload', () => {
   describe('constructor', () => {
