+++ packages/utils/tsconfig.json
@@ -1,8 +1,9 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig.json",
   "extends": "../../tsconfig.base.json",
-  "include": ["src", "test"],
+  "include": ["src"],
   "compilerOptions": {
+    "rootDir": "src",
     "outDir": "dist",
   }
 }
